+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/utils/context/AuthTokenInfo.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-package com.microsoft.appcenter.utils.storage;
+package com.microsoft.appcenter.utils.context;
 
 import java.util.Date;
 
@@ -12,20 +12,20 @@ import java.util.Date;
  */
 public class AuthTokenInfo {
 
-    private final String mToken;
+    private final String mAuthToken;
 
     private final Date mStartTime;
 
     private final Date mEndTime;
 
-    public AuthTokenInfo(String token, Date startTime, Date endTime) {
-        mToken = token;
+    public AuthTokenInfo(String authToken, Date startTime, Date endTime) {
+        mAuthToken = authToken;
         mStartTime = startTime;
         mEndTime = endTime;
     }
 
-    public String getToken() {
-        return mToken;
+    public String getAuthToken() {
+        return mAuthToken;
     }
 
     public Date getStartTime() {
